refactor(VTID-02807): extract orb-tool dispatcher to shared module (PR A of 3)
PR A of the lift-not-duplicate refactor that addresses the user's
concern: 'why do you have to fix this when we said we are mirroring
Vertex?'

Background: services/gateway/src/routes/orb-tool.ts began as a parallel
re-implementation of the inline tool-dispatch logic that lives inside
orb-live.ts (Vertex pipeline). 22 tools each got a hand-written handler
that immediately drifted from Vertex behaviour:
  - search_events lost titles in the LLM-facing text response
  - search_community queries community_groups (table doesn't exist)
  - and 20 others unverified.

Each drift required a separate fix; with 22 tools at risk this compounds
indefinitely.

PR A (this PR): pure structural refactor, zero behaviour change.
  - Move all 22 tool_* implementations from orb-tool.ts into a new
    services/gateway/src/services/orb-tools-shared.ts module.
  - Export ORB_TOOL_REGISTRY + dispatchOrbTool() so any caller (LiveKit
    HTTP wrapper, Vertex inline handler, future test harness) routes
    through the same code path.
  - orb-tool.ts becomes a 60-line HTTP wrapper that resolves auth +
    Supabase admin client and delegates to dispatchOrbTool.

PR B (next): port Vertex's inline tool-cases in orb-live.ts to call
dispatchOrbTool. Each Vertex case keeps its current arg shape but its
body becomes 'return await dispatchOrbTool(...)'. Drift becomes
impossible because both pipelines run identical code.

PR C (after): parity-scanner CI that walks both call sites and fails
the build if any tool name diverges from the registry.

Build verified locally with npm run build.
